refactor(croquis-cf): share module path candidates
