+++ crates/vize_croquis_cf/src/lib.rs
@@ -54,6 +54,7 @@
 mod analyzer;
 mod diagnostics;
 mod graph;
+mod module_paths;
 mod registry;
 mod suppression;
